requirements.txt
astroid==2.0.4
certifi==2018.8.24
colorama==0.4.3
dj-database-url==0.5.0
Django==2.2.15
django-cors-headers==3.4.0
djangorestframework==3.11.1
gunicorn==20.0.4
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pylint==2.1.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
typed-ast==1.1.0
whitenoise==5.2.0
wincertstore==0.2
wrapt==1.10.11
